Added budget progress feature.

+++ src/main/java/com/aks/finance/tracker/beans/TransactionResponseBean.java
@@ -28,7 +28,7 @@ public class TransactionResponseBean {
 
     private String transactionCode;
 
-    private float amount;
+    private Double amount;
 
     private Long id;
